requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
pillow==11.1.0
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
